requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
jsonschema==3.2.0
pyrsistent==0.18.0
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
